Changed Id type from Integer to Long and fixed bug in 'Event Calendar' for Patient

+++ src/main/java/mk/ukim/finki/wpproekt/model/Bolnica.java
@@ -15,17 +15,17 @@ import javax.persistence.Id;
 public class Bolnica {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer bolnica_id;
+    private Long bolnica_id;
     private String naziv;
     private String grad;
-    private Integer broj;
+    private Long broj;
     private String ulica;
     private String smetka_bolnica;
 
     public Bolnica() {
     }
 
-    public Bolnica(String naziv, String grad, Integer broj, String ulica, String smetka_bolnica) {
+    public Bolnica(String naziv, String grad, Long broj, String ulica, String smetka_bolnica) {
         this.naziv = naziv;
         this.grad = grad;
         this.broj = broj;
