+++ src/main/java/mk/ukim/finki/wpproekt/model/exceptions/OddelNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(code = HttpStatus.NOT_FOUND)
 public class OddelNotFoundException extends RuntimeException {
-    public OddelNotFoundException(Integer id) {
+    public OddelNotFoundException(Long id) {
         super(String.format("Department with id: %d is not found", id));
 
     }
