+++ src/main/java/mk/ukim/finki/wpproekt/repository/LekoviRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LekoviRepository extends JpaRepository<Lekovi, Integer> {
+public interface LekoviRepository extends JpaRepository<Lekovi, Long> {
 }
