+++ src/main/java/mk/ukim/finki/wpproekt/repository/TransakcijaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransakcijaRepository extends JpaRepository<Transakcija, Integer> {
+public interface TransakcijaRepository extends JpaRepository<Transakcija, Long> {
 }
