+++ src/main/java/mk/ukim/finki/wpproekt/repository/UpatRepository.java
@@ -9,8 +9,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UpatRepository extends JpaRepository<Upat, Integer> {
-    Optional<Upat> findById(Integer upat_id);
+public interface UpatRepository extends JpaRepository<Upat, Long> {
+    Optional<Upat> findById(Long upat_id);
 
     @Query("select u from Upat u left join Rezervacija r on u.upat_id = r.upat.upat_id left join Transakcija t on r.rezervacija_id = t.rezervacija.rezervacija_id where (r.upat.upat_id is null) or (r.upat.upat_id is not null and t.rezervacija.rezervacija_id is null)")
     List<Upat> findAllWithoutReservation();
