+++ src/main/java/mk/ukim/finki/wpproekt/sevice/PrepisaniLekoviService.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface PrepisaniLekoviService {
 
     void save (PrepisaniLekovi prepisaniLekovi);
-    PrepisaniLekovi save (String doktor_username, Integer lek_id, Integer pacient_id);
+    PrepisaniLekovi save (String doktor_username, Long lek_id, Long pacient_id);
     List<Lekovi> findLekoviZaPacient();
     List<PrepisaniLekovi> findAll();
-    PrepisaniLekovi save (String doktor_username, Integer lek_id, String pacient_username);
+    PrepisaniLekovi save (String doktor_username, Long lek_id, String pacient_username);
 }
