+++ src/main/java/mk/ukim/finki/wpproekt/sevice/TerminService.java
@@ -9,11 +9,11 @@ import java.util.Optional;
 
 public interface TerminService {
     Optional<Termin> findTerminById (TerminId terminId);
-    List<Termin> findTerminByTermin_id (Integer termin_id);
-    Termin findOneTerminByTerminId (Integer termin_id);
+    List<Termin> findTerminByTermin_id (Long termin_id);
+    Termin findOneTerminByTerminId (Long termin_id);
     void save (Termin termin);
     void deleteTermin (Termin termin);
     List<Termin> findAll();
     List<Termin> findOnlyFutureAndFree (ZonedDateTime now);
-    List<Termin> findOnlyFutureAndFreeAndByDoktor (ZonedDateTime now, Integer doktor_id);
+    List<Termin> findOnlyFutureAndFreeAndByDoktor (ZonedDateTime now, Long doktor_id);
 }
